+++ plugins/11_TheWarWithin/localization/deDE.lua
@@ -631,3 +631,6 @@ L['options_icons_youre_my_friend_now'] = '{achievement:41708}'
 L['options_icons_youre_my_friend_now_desc'] = 'Zeigt die Positionen von {npc:239248} für den Erfolg {achievement:41708} an.'
 L['rat_in_building_rail'] = 'Im Gebäude auf dem Geländer.'
 L['rat_2nd_floor_trash_can'] = 'Im Gebäude im zweiten Stock auf einem Mülleimer.'
+
+L['sewer_cheese_label'] = nil
+L['undermine_undershirt_note'] = nil