+++ plugins/11_TheWarWithin/localization/esES.lua
@@ -627,3 +627,6 @@ L['options_icons_youre_my_friend_now'] = '{achievement:41708}'
 L['options_icons_youre_my_friend_now_desc'] = 'Muestra dónde están las {npc:239248} para el logro {achievement:41708}.'
 L['rat_in_building_rail'] = 'En el riel dentro del edificio.'
 L['rat_2nd_floor_trash_can'] = 'En la papelera en el segundo piso dentro del edificio.'
+
+L['sewer_cheese_label'] = nil
+L['undermine_undershirt_note'] = nil