+++ plugins/11_TheWarWithin/localization/ruRU.lua
@@ -628,3 +628,6 @@ L['options_icons_youre_my_friend_now'] = '{achievement:41708}'
 L['options_icons_youre_my_friend_now_desc'] = 'Показать {npc:239248} места для {achievement:41708}.'
 L['rat_in_building_rail'] = 'Внутри здания на рельсах.'
 L['rat_2nd_floor_trash_can'] = 'Внутри здания на втором этаже на мусорном баке.'
+
+L['sewer_cheese_label'] = nil
+L['undermine_undershirt_note'] = nil