+++ plugins/11_TheWarWithin/localization/zhCN.lua
@@ -627,3 +627,6 @@ L['options_icons_youre_my_friend_now'] = '{achievement:41708}'
 L['options_icons_youre_my_friend_now_desc'] = '显示 {achievement:41708} 成就中 {npc:239248} 的位置。'
 L['rat_in_building_rail'] = '在建筑物内的栏杆上。'
 L['rat_2nd_floor_trash_can'] = '在建筑物二楼垃圾桶上。'
+
+L['sewer_cheese_label'] = nil
+L['undermine_undershirt_note'] = nil