+++ plugins/11_TheWarWithin/localization/zhTW.lua
@@ -627,3 +627,6 @@ L['options_icons_youre_my_friend_now'] = '{achievement:41708}'
 L['options_icons_youre_my_friend_now_desc'] = nil
 L['rat_in_building_rail'] = nil
 L['rat_2nd_floor_trash_can'] = nil
+
+L['sewer_cheese_label'] = nil
+L['undermine_undershirt_note'] = nil